backend/requirements.txt
# WhatsNaija AI Backend Dependencies

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP client for API calls
httpx==0.27.0

# Groq API
groq==0.4.2

# Database
supabase==2.10.0

# Environment variables
python-dotenv==1.0.0

# Data validation (updated for Python 3.13 compatibility)
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Security
python-multipart==0.0.6
PyJWT==2.8.0
